Move realm/vault initialization to application (android)
